requirements.txt
Flask>=2.1
Flask_SQLAlchemy
psycopg2-binary
Werkzeug==2.0.3
